pyproject.toml
[tool.poetry]
name = "benchmark_influx_db"
version = "0.1.0"
description = ""
authors = ["Karl <karlwolffgang@googlemail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
gewv-timeseries-client = "^0.5.0"
python-dotenv = "^0.19.0"
loguru = "^0.5.3"
certifi = "^2021.5.30"

[tool.poetry.dev-dependencies]
black = {version = "^21.9b0", allow-prereleases = true}
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
